find one place with size_t and int problem
git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codesv3@4502 eb38b3b0-33ea-0310-94ad-d43fea0d214f

+++ ON/ON-NEGF-share/GetNlop_on.cpp
@@ -42,7 +42,7 @@ void GetNlop_on(void)
     int ion, idx, ip;
     int tot_prj, ion1, index;
     size_t PROJECTOR_SPACE;
-    int prjcount;
+    size_t prjcount;
     double *beta;
     SPECIES *sp;
     ION *iptr;
